feat: Add periods management and enhance competencies system
- Add PeriodosController with full CRUD operations
- Create PeriodosForm for period management
- Add periods view with DataTables integration
- Enhance competencies editing with percentage adjustment service
- Update materias controller with improved form handling
- Refactor grading services for better performance
- Update database migrations for periods and final grades
- Add navigation menu items for new features
- Integrate DataTables JS library for improved UI
- Update dashboard and layout components
- Add progress bar component enhancements

+++ database/migrations/2025_06_09_202814_create_notas_finales_competencias_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('estudiante_id');
             $table->unsignedBigInteger('materia_id');
             $table->unsignedBigInteger('competencia_id');
-            $table->double('nota_final')->default(0);
+            $table->decimal('nota_final', 8, 2)->default(0);
             $table->timestamps();
 
             $table->foreign('estudiante_id')->references('id')->on('usuarios')->onDelete('cascade');
